Move PLRoute to redirect-to-english-page.tsx.
