+++ frontend/lib/pages/redirect-to-english-page.tsx
@@ -1,10 +1,16 @@
-import React from "react";
+import React, { useContext } from "react";
 import Page from "../ui/page";
 import { li18n } from "../i18n-lingui";
 import { t, Trans } from "@lingui/macro";
 import * as H from "history";
-import { useHistory, RouteComponentProps } from "react-router-dom";
-import i18n from "../i18n";
+import {
+  useHistory,
+  RouteComponentProps,
+  Route,
+  RouteProps,
+} from "react-router-dom";
+import i18n, { SupportedLocale } from "../i18n";
+import { AppContext } from "../app-context";
 
 export const RedirectToEnglishPage: React.FC<{ to: H.LocationDescriptor }> = (
   props
@@ -35,3 +41,51 @@ export const RedirectCurrentPathToEnglishPage: React.FC<RouteComponentProps> = (
     throw new Error(`Path ${location.pathname} is not locale-prefixed!`);
   return <RedirectToEnglishPage to={{ ...location, pathname }} />;
 };
+
+type ExtraPLRouteProps = {
+  locales?: SupportedLocale[];
+  wipLocales?: SupportedLocale[];
+};
+
+type PLRouteProps = RouteProps & ExtraPLRouteProps;
+
+function isLocaleSupported(
+  options: ExtraPLRouteProps & { enableWipLocales: boolean }
+): boolean {
+  let locales = options.locales || ["en"];
+
+  if (options.enableWipLocales && options.wipLocales) {
+    locales = [...locales, ...options.wipLocales];
+  }
+
+  return new Set(locales).has(i18n.locale);
+}
+
+/** A partially localized route. */
+export const PLRoute: React.FC<PLRouteProps> = (props) => {
+  const { server } = useContext(AppContext);
+
+  if (isLocaleSupported({ ...server, ...props })) {
+    return <Route {...props} />;
+  }
+
+  return (
+    <Route
+      {...props}
+      children={undefined}
+      render={undefined}
+      component={RedirectCurrentPathToEnglishPage}
+    />
+  );
+};
+
+export function toPLRoute(
+  el: JSX.Element,
+  plProps?: ExtraPLRouteProps
+): JSX.Element {
+  if (process.env.NODE_ENV !== "production" && el.type !== Route) {
+    console.warn("toPLRoute() expected a <Route> JSX Element!");
+  }
+  const routeProps: RouteProps = el.props;
+  return <PLRoute key={el.key ?? undefined} {...routeProps} {...plProps} />;
+}
